refacto server architecture
